part-3-evaluations/solution/requirements.txt
strands-agents
strands-agents-tools
strands-agents-evals[simulators]
